AppDir/usr/share/darktable/mcp/requirements.txt
opencv-python-headless
numpy
rembg
requests
pillow
